backend/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.7.2
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
Django==4.2.2
django-apscheduler==0.6.2
django-cors-headers==4.0.0
django-environ==0.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.7.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
